requirements.txt
pandas
openpyxl
python-docx
pyautogui
keyboard
Pillow